+++ CHANGELOG.md
@@ -31,6 +31,10 @@ $form-field-transition:                background-color $default-transition-dura
 
 ### NEW
 
+- Softpage: New event `closeSoftpage` to close softpage from a custom toggle:
+```js
+$(window).trigger('closeSoftpage');
+```
 - Softpage: Option for loading content of a specific element instead of href-attribute added:
 ```HTML
 <a href="/link-to-page" data-trigger-softpage data-softpage-content-id="example-element">I will trigger a softpage</a>